feat(collections): extract manageCollection into @mulmoclaude/core/collection/server
The agent's collection data plane (getItems with computed fields,
schema-validated putItems, getOntology, schemaDocs, getSchema/putSchema)
moves from the host's mcp-tools into the shared core so MulmoTerminal
can expose the same tool instead of leaving its agent on raw file I/O.

Host specifics ride ManageCollectionDeps, extended by two injections:
- bundledHelpsDir: workspace-setup's helpsAssetDir is ESM-only
  (import.meta.url, build pass 2) while collection/server builds dual
  ESM+CJS, so the schemaDocs fallback dir is injected; the host shim's
  re-exported factory pre-binds it, keeping the test contract.
- refreshAfterWrite has no in-package default anymore: the post-putSchema
  refresh (scheduled skills + user tasks) is MulmoClaude-side state, so
  the shim's singleton binds it and the package default is no refresh.

The shim keeps every prior export (manageCollection singleton with the
ablation env binding, makeManageCollectionTool, MAX_* constants), so
routes and tests are untouched. Core 0.18.1 -> 0.19.0; collection-plugin
widens its core range to "^0.18.0 || ^0.19.0" (0.11.0 -> 0.11.1, no code
change) so a registry install against core 0.19 satisfies peers.
