+++ packages/mulmoclaude/package.json
@@ -37,7 +37,7 @@
     "@mulmoclaude/accounting-plugin": "^0.3.2",
     "@mulmoclaude/chart-plugin": "^0.1.3",
     "@mulmoclaude/collection-plugin": "^0.11.0",
-    "@mulmoclaude/core": "^0.18.1",
+    "@mulmoclaude/core": "^0.19.0",
     "@mulmoclaude/form-plugin": "^0.1.4",
     "@mulmoclaude/html-plugin": "^0.2.5",
     "@mulmoclaude/markdown-plugin": "^0.1.10",
